final cleanup

+++ services/storage/src/simcore_service_storage/simcore_s3_dsm.py
@@ -276,15 +276,16 @@ class SimcoreS3DataManager(BaseDataManager):
                 conn, parse_obj_as(SimcoreS3FileID, file_id)
             )
 
-        # ensure file is deleted first in case it already exists
-        await self.delete_file(
-            user_id=user_id,
-            file_id=file_id,
-            # NOTE: bypassing check since the project access rights don't play well
-            # with collaborators
-            # SEE https://github.com/ITISFoundation/osparc-simcore/issues/5159
-            enforce_access_rights=False,
-        )
+        if not is_directory:
+            # ensure file is deleted first in case it already exists
+            await self.delete_file(  # NOTE: MD check
+                user_id=user_id,
+                file_id=file_id,
+                # NOTE: bypassing check since the project access rights don't play well
+                # with collaborators
+                # SEE https://github.com/ITISFoundation/osparc-simcore/issues/5159
+                enforce_access_rights=False,
+            )
         async with self.engine.acquire() as conn:
             # initiate the file meta data table
             fmd = await self._create_fmd_for_upload(
